HTML5 Part 1
I've ensured the Stacey CMS install is at 2.3.0. I've replaced the old
markup with HTML5 tags. I've stripped down the stylesheet to a minimum
to preserve readability but that's it. I've added header and footer
partials and included them in the templates (except for article, which
won't take the header partial for no good reason).